Added book images
I added Book images as a string to the book object as Base64 encoded
images. Also the adding of books will add book images as well as
all the other previous information that was initally missing. This
will be added as a blob in the database. The servlet will then
take the image as a base 64 in order to display it to the user